test3/urls/requirements.txt
aiohttp>=3.9
aiodns>=3.1
beautifulsoup4>=4.12
click>=8.1
dnspython>=2.4
fastapi>=0.110
geoip2>=4.7
Levenshtein>=0.25
pyOpenSSL>=24.0
python-dotenv>=1.0.0
python-whois>=0.9
requests>=2.31
tldextract>=5.1
uvicorn>=0.28
validators>=0.28